Added template for PowerShell in Lambda functions
Took the code I wrote for an old PR in the aws-sam-cli repo
(https://github.com/aws/aws-sam-cli/pull/820) and freshened it up for
this templates library. I have put them under .net as that's the runtime
as far as Lambda is concerned, and have had to use the makefile build
approach to get it to properly build out the dependencies (since letting
the default dotnet one run wouldn't yield any results in this case).